+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - **react** - Allow globally configuring the React Query key via `configureQueryKey`. Export `DEFAULT_QUERY_KEY`.
+- **core** - Add a query builder.
 
 ## [1.0.0-alpha.5] - 2025-12-30