+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.0-alpha.6] - 2026-01-10
+
 ### Added
 
 - **react** - Allow globally configuring the React Query key via `configureQueryKey`. Export `DEFAULT_QUERY_KEY`.